Don't make unnecessary network requests
Closes #54
